agent: Rewrite custom field policy related code
