+++ agent/crates/public/src/lib.rs
@@ -18,6 +18,7 @@ pub mod bitmap;
 pub mod buffer;
 pub mod bytes;
 pub mod chrono_map;
+pub mod codecs;
 pub mod consts;
 pub mod counter;
 pub mod debug;
