+++ agent/src/trident.rs
@@ -975,6 +975,7 @@ impl Trident {
                         let callbacks = config_handler.on_config(
                             cfg.user_config,
                             &exception_handler,
+                            &stats_collector,
                             None,
                             #[cfg(target_os = "linux")]
                             &api_watcher,
@@ -1055,6 +1056,7 @@ impl Trident {
                     let callbacks = config_handler.on_config(
                         user_config,
                         &exception_handler,
+                        &stats_collector,
                         None,
                         #[cfg(target_os = "linux")]
                         &api_watcher,
@@ -1119,6 +1121,7 @@ impl Trident {
                         .on_config(
                             user_config,
                             &exception_handler,
+                            &stats_collector,
                             Some(components),
                             #[cfg(target_os = "linux")]
                             &api_watcher,
@@ -1167,6 +1170,7 @@ impl Trident {
                     config_handler.on_config(
                         user_config,
                         &exception_handler,
+                        &stats_collector,
                         None,
                         #[cfg(target_os = "linux")]
                         &api_watcher,
